+++ tests/test_repeat.ts
@@ -1982,9 +1982,7 @@ describe('repeat', function () {
       queueEvents.on('waiting', function ({ jobId }) {
         try {
           expect(jobId).to.be.equal(
-            `repeat:16db7a9b166154f5c636abf3c8fe3364:${
-              date.getTime() + 1 * ONE_SECOND
-            }`,
+            `repeat:16db7a9b166154f5c636abf3c8fe3364:${date.getTime() + 1 * ONE_SECOND}`,
           );
           resolve();
         } catch (err) {